requirements.txt
# Core dependencies
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
pysrt>=1.1.2
webvtt-py>=0.4.6
click>=8.1.0
PyYAML>=6.0.0
python-dateutil>=2.8.0

# AI/ML dependencies
openai-whisper>=20250625
anthropic>=0.28.0
torch>=2.0.0

# Note: MediaPipe not available for Python 3.13 yet
# Using OpenCV Haar cascades for face detection instead